+++ CHANGELOG.rst
@@ -6,8 +6,8 @@ All notable changes to this project will be documented in this file.
 The format is based on `Keep a
 Changelog <https://keepachangelog.com/en/1.1.0/>`__.
 
-Unreleased
-----------
+2.2.0 - 2026-05-26
+------------------
 
 Added
 ~~~~~